refactor[tcg]: separated items into one file per category by function
